Refactor some class access modifiers
