+++ src/main/java/org/mvplugins/multiverse/inventories/commands/prompts/GroupEditPrompt.java
@@ -12,7 +12,7 @@ import org.mvplugins.multiverse.inventories.util.MVInvi18n;
 
 import static org.mvplugins.multiverse.core.locale.message.MessageReplacement.replace;
 
-class GroupEditPrompt extends InventoriesPrompt {
+final class GroupEditPrompt extends InventoriesPrompt {
 
     public GroupEditPrompt(final MultiverseInventories plugin, final MVCommandIssuer issuer) {
         super(plugin, issuer);
