+++ src/main/java/org/mvplugins/multiverse/inventories/event/WorldChangeShareHandlingEvent.java
@@ -5,7 +5,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
 
-public class WorldChangeShareHandlingEvent extends ShareHandlingEvent implements Cancellable {
+public final class WorldChangeShareHandlingEvent extends ShareHandlingEvent implements Cancellable {
 
     private static final HandlerList HANDLERS = new HandlerList();
