+++ src/main/java/org/mvplugins/multiverse/inventories/util/ItemStackConverter.java
@@ -55,6 +55,6 @@ public final class ItemStackConverter {
     }
 
     private ItemStackConverter() {
-        // no instantiation
+        throw new IllegalStateException();
     }
 }
